losses are exploding

+++ video_vae/train/main.py
@@ -51,7 +51,7 @@ def main(args):
           p.requires_grad = False 
 
       optimizer_g.zero_grad()
-      with torch.autocast(device_type="cuda", dtype=torch.float16):
+      with torch.autocast(device_type="cuda", dtype=torch.float32):
 
         
         posterior, reconstruct = vae(train_video_dataloaders)
@@ -82,7 +82,7 @@ def main(args):
           p.requires_grad = True
 
       optimizer_d.zero_grad()
-      with torch.autocast(device_type="cuda", dtype=torch.float16):
+      with torch.autocast(device_type="cuda", dtype=torch.float32):
 
         
         gan_loss, gan_log  = loss(train_video_dataloaders,
